+++ scene/world.hpp
@@ -119,7 +119,7 @@ public:
         entity->attach_component(transform);
         RenderComponent render;
         render.set_mesh(MeshPrimitive::Sphere);
-        render.pbr_scalar_parameters = Vec3f(0.0, 1.0 / (double)(SIZE - 1) * i, (1.0 / (double)(SIZE - 1)) * j);
+        render.pbr_scalar_parameters = Vec3f(0.0, 1.0f / (float)(SIZE - 1) * i, (1.0f / (float)(SIZE - 1)) * j);
         render.set_shading_model(ShadingModel::PhysicallyBasedScalars);
         entity->attach_component(render);
         ActionComponent action([=](uint64_t frame, uint64_t dt) {
